added tests to all my calc methods

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -18,7 +18,7 @@ public class MyMathCalc {
     }
 
 
-    public double divide (double a, double b) {
+    public double divide (int a, int b) {
         return a/b;
 }
 }
